YG-76 📂 #open #comment 폴더 구조 변경
