+++ app/(afterLogin)/createPost/_components/calendar/type.ts
@@ -1,0 +1,9 @@
+export type selectCalendarProps = {
+    isOpen: boolean
+    onClick: () => void
+}
+
+export type DateRange = {
+    start: Date | null
+    end: Date | null
+}
