+++ app/(afterLogin)/createPost/_components/form/type.ts
@@ -1,0 +1,17 @@
+import { createPostTemplate } from "@/apis/type"
+
+export type FormInputsProps = {
+    formText: string
+    formData: createPostTemplate
+    handleInputChange: <K extends keyof createPostTemplate>(field: K, value: createPostTemplate[K]) => void
+}
+
+export type FormSelectorProps = {
+    onClick: () => void
+    label: string
+    state: "continent" | "calendar"
+}
+
+export type FormBtnProps = {
+    setIsOverlayOpen: (isOpen: boolean) => void
+}
