+++ app/(afterLogin)/createPost/_components/overlay/mapOverlay.tsx
@@ -1,0 +1,4 @@
+const MyMapOverlay = () => {
+    return <div>this is my map overlay</div>
+}
+export default MyMapOverlay
