+++ app/(afterLogin)/createPost/_components/overlay/thumbnailUploader.tsx
