+++ app/(afterLogin)/detailPost/[postId]/_components/comment/likeToComment.tsx
