+++ app/(afterLogin)/detailPost/[postId]/_components/comment/type.ts
@@ -1,0 +1,42 @@
+export type CommentProps = {
+    postId: number
+}
+
+export type Comment = {
+    id: number
+    content: string
+    nickname: string
+    postId: number
+    createdAt: string
+    modifiedAt: string
+    likeCount: number
+}
+
+export type CommentBoxProps = {
+    comments: Comment[]
+}
+
+export type CommentType = {
+    id: number
+    content: string
+    likes: number
+    comments: number
+    date: string
+    userId: string
+    userProfileImage: string
+}
+
+export type ReCommentType = {
+    id: number
+    content: string
+    likes: number
+    date: string
+    userId: string
+    userProfileImage: string
+    parentId: number
+}
+
+export type LikeToComments = {
+    likes: number
+    comments: number
+}
