+++ app/(afterLogin)/detailPost/[postId]/_components/floating/floatingButton.tsx
