+++ app/(afterLogin)/detailPost/[postId]/_components/floating/type.ts
@@ -1,0 +1,10 @@
+export type FloatingIcon = {
+    name: string
+    icon: string
+    isActive: boolean
+}
+
+export type FloatingButtonType = {
+    icon: FloatingIcon
+    onClick: () => void
+}
