+++ app/(afterLogin)/detailPost/[postId]/_components/type.ts
@@ -1,5 +1,4 @@
 import { ShortPosts } from "@/utils/type"
-import { Comment } from "../type"
 
 export type PostDetailProps = {
     title: string
@@ -9,24 +8,4 @@ export type PostDetailProps = {
     destination: string
     travel_range: string
     shortPosts: ShortPosts[]
-}
-
-export type CommentBoxProps = {
-    comments: Comment[]
-}
-
-export type FloatingIcon = {
-    name: string
-    icon: string
-    isActive: boolean
-}
-
-export type FloatingButtonType = {
-    icon: FloatingIcon
-    onClick: () => void
-}
-
-export type LikeToComments = {
-    likes: number
-    comments: number
 }
