+++ app/(afterLogin)/user/[userId]/_components/profile/profile.tsx
@@ -1,4 +1,4 @@
-import { ProfileProps } from "../type"
+import { ProfileProps } from "./type"
 
 const Profile = ({ name, bio, profileImage, bgImage, onEdit }: ProfileProps) => {
     return (
