+++ app/(afterLogin)/user/[userId]/_components/userDetails.tsx
